factor out typed API module
