+++ src/api/api.ts
@@ -1,42 +1,9 @@
 import { Body, fetch as tauriFetch, ResponseType } from '@tauri-apps/api/http';
 
-import { paths } from './api-paths';
 import { REFSTUDIO_HOST } from './server';
 
 type HttpMethod = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH';
-type LowerMethod = 'get' | 'post' | 'put' | 'delete' | 'patch';
 type ResponseParser = 'JSON' | 'ArrayBuffer';
-
-type Unionize<T> = { [k in keyof T]: { k: k; v: T[k] } }[keyof T];
-type ExtractPathsForMethod<T, Method extends LowerMethod> = T extends { k: infer Path; v: Record<Method, unknown> }
-  ? Path
-  : never;
-type PathsForMethod<Method extends LowerMethod> = ExtractPathsForMethod<Unionize<paths>, Method>;
-
-type ParamsFor<Path extends keyof paths, Method extends LowerMethod> = paths extends Record<
-  Path,
-  Record<Method, { parameters: infer Params }>
->
-  ? Params
-  : never;
-
-type JsonResponseFor<Path extends keyof paths, Method extends LowerMethod> = paths extends Record<
-  Path,
-  Record<Method, { responses: { 200: { content: { 'application/json': infer ResponseType } } } }>
->
-  ? ResponseType
-  : never;
-
-type JsonRequestBodyFor<Path extends keyof paths, Method extends LowerMethod> = paths extends Record<
-  Path,
-  Record<Method, { requestBody: { content: { 'application/json': infer RequestBody } } }>
->
-  ? RequestBody
-  : never;
-
-type PathArgs<Path extends keyof paths, Method extends LowerMethod> = [ParamsFor<Path, Method>] extends [never]
-  ? []
-  : [options: ParamsFor<Path, Method>];
 
 /** Issue a GET request using either web fetch or Tauri fetch */
 export async function universalGet<ResponsePayload = unknown>(
@@ -44,43 +11,6 @@ export async function universalGet<ResponsePayload = unknown>(
   responseParser: ResponseParser = 'JSON',
 ): Promise<ResponsePayload> {
   return universalRequest<ResponsePayload>('GET', path, undefined, responseParser);
-}
-
-interface RouteParameters {
-  path?: Record<string, string>;
-  query?: Record<string, string>;
-}
-
-function completePath(path: string, params: RouteParameters) {
-  if (params.path) {
-    for (const [key, value] of Object.entries(params.path)) {
-      path = path.replace('{' + key + '}', value);
-    }
-  }
-  if (params.query) {
-    path += '?' + new URLSearchParams(params.query).toString();
-  }
-  return path;
-}
-
-export async function apiGetJson<Path extends PathsForMethod<'get'>>(pathSpec: Path, ...args: PathArgs<Path, 'get'>) {
-  const options = (args as unknown[])[0] as RouteParameters | undefined;
-  const path = options ? completePath(pathSpec, options) : pathSpec;
-  console.log('apiGetJSON', pathSpec, path);
-  type ResponseType = JsonResponseFor<Path, 'get'>;
-  return universalRequest<ResponseType>('GET', path, undefined);
-}
-
-export async function apiPost<Path extends PathsForMethod<'post'>>(
-  pathSpec: Path,
-  ...args: [...params: PathArgs<Path, 'post'>, body: JsonRequestBodyFor<Path, 'post'>]
-) {
-  const safeArgs = args as unknown as [params: RouteParameters, body: unknown] | [body: unknown];
-  const [options, body] = safeArgs.length === 2 ? safeArgs : [undefined, ...safeArgs];
-  const path = options ? completePath(pathSpec, options) : pathSpec;
-  console.log('apiPostJSON', pathSpec, path);
-  type ResponseType = JsonResponseFor<Path, 'post'>;
-  return universalPost<ResponseType>(path, body);
 }
 
 /** Issue a POST request with a JSON payload using either web fetch or Tauri fetch */
